requirements.txt
dbus-python==1.3.2
gobject==0.1.0
ioctl==0.2.2
keyboard==0.13.5
pycairo==1.25.1
PyGObject==3.46.0
pyudev==0.24.1
